#1 Working draft of new schema
The basic principle behind this schema is that you define your three
main "entities" at the start -- Holdings (a.k.a. Premises), Persons and
Animals.

You can then create a few different type of transactions which
reference these such as:

 - Sightings (at a single location)
 - Movements (from A to B)

The schema is quite “rich” and I don’t think there’s any need to
support all of it from day 1. But I like that we’d have a pretty
powerful schema that was stable that we could work towards implementing
over time.